+++ examples/data-sources/iosxr_router_isis_interface/data-source.tf
@@ -1,4 +1,0 @@
-data "iosxr_router_isis_interface" "example" {
-  process_id     = "P1"
-  interface_name = "GigabitEthernet0/0/0/1"
-}