+++ examples/data-sources/iosxr_router_isis_net/data-source.tf
@@ -1,4 +1,0 @@
-data "iosxr_router_isis_net" "example" {
-  process_id = "P1"
-  net_id     = "49.0001.2222.2222.2222.00"
-}