+++ internal/provider/resource_iosxr_router_isis.go
@@ -53,6 +53,156 @@ func (t resourceRouterISISType) GetSchema(ctx context.Context) (tfsdk.Schema, di
 					helpers.StringEnumValidator("level-1", "level-1-2", "level-2-only"),
 				},
 			},
+			"nets": {
+				MarkdownDescription: helpers.NewAttributeDescription("A Network Entity Title (NET) for this process").String,
+				Optional:            true,
+				Attributes: tfsdk.ListNestedAttributes(map[string]tfsdk.Attribute{
+					"net_id": {
+						MarkdownDescription: helpers.NewAttributeDescription("A Network Entity Title (NET) for this process").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+					},
+				}, tfsdk.ListNestedAttributesOptions{}),
+			},
+			"address_families": {
+				MarkdownDescription: helpers.NewAttributeDescription("IS-IS address family").String,
+				Optional:            true,
+				Attributes: tfsdk.ListNestedAttributes(map[string]tfsdk.Attribute{
+					"af_name": {
+						MarkdownDescription: helpers.NewAttributeDescription("Address family name").AddStringEnumDescription("ipv4", "ipv6").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.StringEnumValidator("ipv4", "ipv6"),
+						},
+					},
+					"saf_name": {
+						MarkdownDescription: helpers.NewAttributeDescription("Sub address family name").AddStringEnumDescription("multicast", "unicast").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.StringEnumValidator("multicast", "unicast"),
+						},
+					},
+					"mpls_ldp_auto_config": {
+						MarkdownDescription: helpers.NewAttributeDescription("Enable LDP IGP interface auto-configuration").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"metric_style_narrow": {
+						MarkdownDescription: helpers.NewAttributeDescription("Use old style of TLVs with narrow metric").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"metric_style_wide": {
+						MarkdownDescription: helpers.NewAttributeDescription("Use new style of TLVs to carry wider metric").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"metric_style_transition": {
+						MarkdownDescription: helpers.NewAttributeDescription("Send and accept both styles of TLVs during transition").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"router_id_interface_name": {
+						MarkdownDescription: helpers.NewAttributeDescription("Router ID Interface").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.StringPatternValidator(0, 0, `[a-zA-Z0-9.:_/-]+`),
+						},
+					},
+					"router_id_ip_address": {
+						MarkdownDescription: helpers.NewAttributeDescription("Router ID address").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"default_information_originate": {
+						MarkdownDescription: helpers.NewAttributeDescription("Distribute a default route").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+				}, tfsdk.ListNestedAttributesOptions{}),
+			},
+			"interfaces": {
+				MarkdownDescription: helpers.NewAttributeDescription("Enter the IS-IS interface configuration submode").String,
+				Optional:            true,
+				Attributes: tfsdk.ListNestedAttributes(map[string]tfsdk.Attribute{
+					"interface_name": {
+						MarkdownDescription: helpers.NewAttributeDescription("Enter the IS-IS interface configuration submode").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.StringPatternValidator(0, 0, `[a-zA-Z0-9.:_/-]+`),
+						},
+					},
+					"circuit_type": {
+						MarkdownDescription: helpers.NewAttributeDescription("Configure circuit type for interface").AddStringEnumDescription("level-1", "level-1-2", "level-2-only").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.StringEnumValidator("level-1", "level-1-2", "level-2-only"),
+						},
+					},
+					"hello_padding_disable": {
+						MarkdownDescription: helpers.NewAttributeDescription("Disable hello-padding").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"hello_padding_sometimes": {
+						MarkdownDescription: helpers.NewAttributeDescription("Enable hello-padding during adjacency formation only").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"priority": {
+						MarkdownDescription: helpers.NewAttributeDescription("Set priority for Designated Router election").AddIntegerRangeDescription(0, 127).String,
+						Type:                types.Int64Type,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.IntegerRangeValidator(0, 127),
+						},
+					},
+					"point_to_point": {
+						MarkdownDescription: helpers.NewAttributeDescription("Treat active LAN interface as point-to-point").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"passive": {
+						MarkdownDescription: helpers.NewAttributeDescription("Do not establish adjacencies over this interface").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"suppressed": {
+						MarkdownDescription: helpers.NewAttributeDescription("Do not advertise connected prefixes of this interface").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"shutdown": {
+						MarkdownDescription: helpers.NewAttributeDescription("Shutdown IS-IS on this interface").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+				}, tfsdk.ListNestedAttributesOptions{}),
+			},
 		},
 	}, nil
 }